Guides for private app and public app
