+++ guides/private-app/.env
@@ -1,1 +1,10 @@
-VITE_CLIENT_ID=XXX
+# API Base URL (copy this from CARTO Workspace -> Developers)
+VITE_API_BASE_URL=https://gcp-us-east1.api.carto.com
+
+# Set this variable if you're using an API Access Token (public application).
+# Go to app.carto.com -> Developers -> Manage API Access Tokens
+# VITE_API_ACCESS_TOKEN=XXX
+
+# Go to app.carto.com -> Developers -> Create new application
+# Set this variable if you're building a private application (it needs to create an application.)
+# VITE_CLIENT_ID=XXX
