+++ guides/private-app/package.json
@@ -1,20 +1,22 @@
 {
-  "name": "carto-auth-spa",
+  "name": "carto-app-template",
   "private": true,
-  "version": "0.0.1",
+  "version": "0.0.0",
   "type": "module",
   "scripts": {
     "dev": "vite --https",
-    "build": "vite build --https",
+    "build": "tsc && vite build --https",
     "preview": "vite preview"
   },
   "devDependencies": {
-    "vite": "^4.3.0"
+    "typescript": "^5.0.2",
+    "vite": "^4.3.2"
   },
   "dependencies": {
-    "@auth0/auth0-spa-js": "^2.0.4",
-    "@deck.gl/carto": "^8.9.9",
+    "@auth0/auth0-spa-js": "^2.0.5",
+    "@deck.gl/carto": "^8.9.16",
+    "@deck.gl/core": "^8.9.16",
     "@vitejs/plugin-basic-ssl": "^1.0.1",
-    "maplibre-gl": "^2.4.0"
+    "maplibre-gl": "^3.0.0"
   }
 }
