+++ guides/private-app/src/style.css
@@ -1,0 +1,46 @@
+body, html {
+  height: 100%;
+  margin: 0;
+  overflow: hidden;
+  padding: 0;
+  position: fixed;
+  width: 100%;
+  font-family: Inter, sans-serif;
+  font-weight: 400;
+  font-size: 1rem;
+}
+#app {
+  flex: 1 1 auto;
+  min-height: 100vh;
+  display: flex;
+  flex-direction: column;
+}
+#map, canvas{
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  max-height: 100%;
+}
+
+header {
+  z-index: 1;
+  position: absolute;
+  padding: 10px;
+}
+
+#login-content {
+  display: none;
+}
+
+.isAuthenticated #login-content {
+  display: block;
+}
+.isAuthenticated #login {
+  display: none;
+}
+
+
+
+
