+++ guides/template/vite.config.js
@@ -1,0 +1,5 @@
+import basicSsl from '@vitejs/plugin-basic-ssl'
+
+export default {
+ plugins: [basicSsl()]
+}
